add method to increase monthly payments and have it affect the graph's results.

+++ db/migrate/20170727151640_create_loans.rb
@@ -3,7 +3,7 @@ class CreateLoans < ActiveRecord::Migration[5.1]
     create_table :loans do |t|
       t.decimal :principal
       t.integer :months
-      t.integer :years
+      t.float :years
       t.decimal :interest
 
       t.timestamps
